+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * re-export `arg` from Cauldron.
 
+* breaking change: `MissingDependenciesError` now includes all the missing dependencies.
+
 ## 0.7.0.0
 
 * Remove dependency on algebraic-graphs, copying those parts of the code that we used.